OpenStarkTFW/CodeRobotFramework/requirements.txt
robotframework 
robotframework-seleniumlibrary
robotframework-appiumlibrary
robotframework-requests 
robotframework-databaselibrary
robotframework-ftplibrary 
robotframework-archivelibrary 
robotframework-difflibrary 
robotframework-mongodbLibrary 
robotframework-httplibrary
unittest-xml-reporting 
html-testRunner 
parameterized 
selenium
requests 
pymysql 
pymongo 
pyautogui 
rabird.winio
https://github.com/BSTester/robotframework-QTLibrary/archive/master.zip